+++ assets/src/scripts/yptSearch.js
@@ -33,7 +33,6 @@ const getCheckedSongFilters = ( inputName ) => {
 	let value = [];
 	const inputs = document.querySelectorAll(`#ypt-ajax-filter-search input[name="${inputName}"]:checked`);
 	Array.prototype.forEach.call(inputs, (element) => {
-		console.log('the element', element.id)
 		if ('radio' === element.type) {
 			value = element.value;
 		} else if ('checkbox' === element.type && "contains-one-barre-chord" !== element.id) {
@@ -61,7 +60,6 @@ const sendAjaxRequest = ( yptSearchResultsElement, searchFormData ) => {
 			url : YPTSEARCHAJAX.ajax_url,
 			data : searchFormData,
 			success : (response) => {
-				console.log(" the response is", response);
 				yptSearchResultsElement.empty();
 				yptSearchResultsCountElement.each(function() {
 					$(this).text(response.data.length);
@@ -223,11 +221,9 @@ const searchFiltersMobile = () => {
 	const yptSearchBlock = $("#ypt-ajax-filter-search");
 	const yptSearchFiltersForm = yptSearchBlock.find("form");
 	const yptSearchResultsElement = yptSearchBlock.find("#ypt-ajax-search-results");
-	console.log('hello')
 
 	yptSearchFiltersForm.on('submit',(e) => {
 		const searchFormData = verifyAndReturnSearchFormData(yptSearchFiltersForm, songFilterCheckboxes);
-		console.log('form data', searchFormData)
 		e.preventDefault();
 		sendAjaxRequest(yptSearchResultsElement, searchFormData);
 	});