+++ header.php
@@ -27,7 +27,8 @@ if ( has_blocks( $post->post_content ) ) {
 	<meta name="viewport" content="width=device-width, initial-scale=1">
 	<link rel="profile" href="http://gmpg.org/xfn/11">
 	<link rel="pingback" href="<?php bloginfo( 'pingback_url' ); ?>">
-
+	<link rel="stylesheet" type="text/css" href="slick/slick.css"/>
+	<link rel="stylesheet" type="text/css" href="slick/slick-theme.css"/>
 	<?php wp_head(); ?>
 </head>
 <body <?php body_class(); ?>>